backend/requirements.txt
# Backend dependencies
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
gunicorn==21.2.0